mcf: avoid intermediate allocations in `push_displayable`
Writes directly to the underlying string, validating the added text
comprises a valid `Field`, and truncating back to the original length in
the event of an error
