+++ mcf/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mcf"
-version = "0.6.0-rc.0"
+version = "0.6.0-rc.1"
 authors = ["RustCrypto Developers"]
 edition = "2024"
 rust-version = "1.85"
